Contribute: Add name field to card form
Add name field to the credit card form, though
it's not required for Stripe token creation,
it's a nice thing to have in the Stripe database,
and it makes things slightly more personal, maybe.

Also change the stylings on the Bootstrap-blue
borders when the focus is on input fields, as
they didn't really fit our palette.