Edit: Work around login issue with TopLevelCtrl
We're failing to work well when our server is
reset, and thereby ending our session. This is
not the end of the world, but it's on our list
of things to fix. In the mean time, adjust the
experience on the Edit page so it's easy to tell
what is happening.