Hello: Limit number of things shown
Set a limit for the number of contributions
to show on the Hello page, because that is
probably less overwelming (and more impactful)
than showing a list of 20 things that a person
might want.

I picked 4, because it seems like a reasonable
number of things to easily keep in one's head
while making a choice, and therefore would
be the best intersection between options
and stress.