Entrance: View the profiles specified
Set the entrance page to show the profiles
of the people specified in config.entranceUsernames.

A 'profile' is a photo, username, real name, and the
communities a person belongs to.

This design is meant for about two profiles, with a
profile-image width of about 270px.

The profile image is retrieved by /profile/:username/image,
which gets an attachment named "profile.jpg" from the CouchDB
doc associated with that :username.