Auth: Save patron to database on first sign-in
When someone signs in the for the first time,
save them to the database. This makes life
a lot easier for the time being, as passport
can then find the profile without caching around.

If this becomes a space issue, or we get a bunch
of empty profiles, well, we should consider
changing the user experience so we don't get as
many abandoned contributions, or see what's going on.