Database: Use nano sometimes
Hack in nano as we begin to try things
on production systems and discover what
is slow or doesn't work.

Nano appears to be slightly faster over
time than cradle, based on a few ad hoc tests.

It's not too much work to have nano and
cradle plugged in at the same time, for the
time being.