Auth: Save user to db (or retrive) when signing in
Move our database operations into passport.serializeUser,
out of passport.deserializeUser, saving the user data
as part of the session.

This allows us to avoid calling the database for every
http request, cutting our response time in half.

The trade-off is that if we need to care when data saved
in the session is changed in the database.