Add the notion of 'members'
Members are people who are members of the site.
That is, they have a profile page, and are able
to receive contributions.

Patrons are not necessarily members. This is mainly
out of fear that it will be a lot of work to maintain
a large member base, at the moment. Ideally, everyone
is a member.