Version 0.6
The database can be on a different computer,
and we can connect via https.

We can send money to other accounts via
Stripe Connect.

There is a rudimentary system for permissions
and roles, meaning we can limit who can create
a profile.

This version might even work, more or less.