UX: Hello: Make frequency choice more visual
Choosing the frequency was a little abstract,
and not straight-forward in usability testing.

Now completely hide the frequency choice if
!thing.canHaz, and default to "once" (for now;
will probably default to "monthly" soon).

Also, remove the 'select' element and use a simple,
see-all-4-options approach.

Lastly, the large black swath for selected
contributions was a bit intense, so the palette
was lightened.