UX: Hello: Further visibility refinements
Hide the contribute button until something
is selected, and make the button more visible
by turning it black.

Introduces a usability issue in that there
is now no way to discontinue contributions;
however, the way of doing that was obtuse,
and needed to be re-thought anyway.

Also clean up the layout slightly of the
frequency hearts.