UX: Hello: Hide contribute panel until called upon
The hello page has a lot of information on it.
Focus on the member, and when the patron is ready,
she can click on "Contribute" if that's what she
wants.

Then we can show the contribute panel.

The support and backers panels have been hidden until
they can be appropriately integrated into the page.