UX: Hello: Clean up header, footer
In the header, show a welcome message,
with a link to the about page, instead
of a cold "anonymous." Also fix the
alignment.

In the footer, remove everything, because
it's distracting on the hello page.

Fix the issue where the footer was floating
in the middle of the page, which was relating
to setting various things to 100% height.