Fix: Edit: Relax on page load
On page load, all of our data watchers fire,
but we don't care, because the data is just
populating from the database.

So, instead of freaking out and trying to
save this "new" data to the database all at
once, resulting in document conflicts because
three updates are firing at the same time,
just relax and put the data into the form.