UX: Edit: Add tabs for easier access to things
The experience of scrolling down a lot to edit
the list of things was not good. Added a tab
menu to make it easier to get to the things.

This introduces a look-and-feel conflict between
the top header and this new tab menu, which we'll
want to look at, but not at this time.