Hello: Support: Show goals progress
We want at least to show progress on our goals
in the support section. Chose to show goals as
percentage complete, rather than absolute dollars
received.

How to show non-goal items is still in need of a
design, but there's a space for it now.

We might decide to show things in a different way
in the future, but I think this will do for now.